requirements.txt
# Flask et dépendances web
Flask==3.0.0
Werkzeug==3.0.1

# Vision et traitement d'image
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# OCR
easyocr>=1.7.0
pytesseract>=0.3.10
requests>=2.32.5

# Note: Tesseract OCR doit être installé sur le système
# Ubuntu/Debian: sudo apt-get install tesseract-ocr tesseract-ocr-ara tesseract-ocr-fra
# macOS: brew install tesseract tesseract-lang
# Windows: https://github.com/UB-Mannheim/tesseract/wiki
